requirements.txt
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
tqdm>=4.66.0